requirements.txt
# python libraries required for this tts system

Flask
python-pptx
edge-TTS
Werkzeug